gap closing run after each scaffolding iteration if >1% gaps

+++ redundans.py
@@ -129,6 +129,18 @@ def run_scaffolding(outdir, scaffoldsFname, fastq, libraries, reducedFname, mapq
             # link output ie out/_sspace.1.1/_sspace.1.1.scaffolds.fasta --> out/_sspace.1.1.scaffolds.fasta
             targetout = os.path.join(os.path.basename(out), os.path.basename(out+".final.scaffolds.fasta"))
             symlink(targetout, pout)
+            # if number of gaps larger than 1%, run gap closer
+            stats     = fasta_stats(open(pout))
+            fastaSize = int(stats.split('\t')[2])
+            gapSize   = int(stats.split('\t')[-2])
+            if 1.0 * gapSize / fastaSize > 0.01:
+                if verbose:
+                    sys.stderr.write("  closing gaps ...\n")
+                nogapsFname = ".".join(pout.split(".")[:-1]) + ".filled.fa"
+                basename = "_sspace.%s.%s._gapcloser"%(i, j)
+                run_gapclosing(outdir, mapq, [libraries[i-1],], nogapsFname, pout, \
+                               threads, limit, iters, 0, basename)
+                pout = nogapsFname            
         # update library insert size estimation, especially for mate-pairs
         libraries = get_libraries(fastq, pout, mapq, threads, limit, verbose=0)
     # create symlink to final scaffolds or pout
@@ -136,12 +148,16 @@ def run_scaffolding(outdir, scaffoldsFname, fastq, libraries, reducedFname, mapq
 
     return libraries
 
-def filter_reads(i, outdir, fq1, fq2, minlen, maxlen, limit, minqual):
+def filter_reads(outdir, fq1, fq2, minlen, maxlen, limit, minqual):
     """Filter FastQ files and return output fnames."""
     fastq = (fq1, fq2)
     # generate output files
-    fn1 = os.path.join(outdir, "_gapcloser.%s.%s.fq"%(i, os.path.basename(fq1.name)))
-    fn2 = os.path.join(outdir, "_gapcloser.%s.%s.fq"%(i, os.path.basename(fq2.name)))
+    fn1 = os.path.join(outdir, "_reads.%s"%(os.path.basename(fq1.name.rstrip('.gz'))))
+    fn2 = os.path.join(outdir, "_reads.%s"%(os.path.basename(fq2.name.rstrip('.gz'))))
+    # skip if fq files already generated
+    if os.path.isfile(fn1) and os.path.isfile(fn2):
+        return fn1, fn2
+    # open output files
     out1 = open(fn1, "w")
     out2 = open(fn2, "w")
     outfiles = (out1, out2, 0, 0)
@@ -170,7 +186,7 @@ def prepare_gapcloser(outdir, mapq, configFn, libFs, libRs, orientations, libIS,
                 sys.stderr.write(info%(orient, fq1, fq2))
             continue
         # filter reads
-        fn1, fn2 = filter_reads(i, outdir, fq1, fq2, minlen, maxlen, limit, mapq)
+        fn1, fn2 = filter_reads(outdir, fq1, fq2, minlen, maxlen, limit, mapq)
         #store config info
         config.append(lines%(iSize, reverse_seq, i, fn1, fn2))
         
@@ -181,77 +197,34 @@ def prepare_gapcloser(outdir, mapq, configFn, libFs, libRs, orientations, libIS,
         return True
     
 def run_gapclosing(outdir, mapq, libraries, nogapsFname, scaffoldsFname, \
-                   threads, limit, iters, verbose, \
+                   threads, limit, iters, verbose, basename="_gapcloser", \
                    overlap=25, maxReadLen=150, minReadLen=40):
     """Execute gapclosing step."""
     pout = scaffoldsFname
     
     for i, (libnames, libFs, libRs, orientations, libIS, libISStDev) in enumerate(libraries, 1):
         # prepare config file and filter reads
-        configFn = os.path.join(outdir, "_gapcloser.%s.conf"%i)
+        configFn = os.path.join(outdir, "%s.%s.conf"%(basename, i))
         # skip if not suitable libraries
         if not prepare_gapcloser(outdir, mapq, configFn, libFs, libRs, orientations, libIS, libISStDev, \
                                  minReadLen, maxReadLen, limit, verbose):
             continue
         # run iterations
-        for j in range(1, iters+1):
-            out = os.path.join(outdir, "_gapcloser.%s.%s.fa"%(i,j))
-            # run GapCloser
-            cmd = ["GapCloser", "-t %s"%threads, "-p %s"%overlap, "-l %s"%maxReadLen, \
-                   "-a", pout, "-b", configFn, "-o", out]
-            if verbose:
-                sys.stderr.write(" iteration %s.%s ...\n"%(i,j))
-                #sys.stderr.write( "  %s\n" % " ".join(cmd) )
-            with open(out+".log", "w") as log:
-                GapCloser = subprocess.Popen(cmd, bufsize=-1, stdout=log, stderr=log)
-                GapCloser.wait()
-            # store out info
-            pout = out
-    # create symlink to final scaffolds or pout
-    symlink(pout, nogapsFname)
-    
-def run_gapclosing2(outdir, mapq, libraries, nogapsFname, scaffoldsFname, \
-                    threads, limit, iters, verbose):
-    """Execute gap closing using Gap2Seq."""
-    pout = scaffoldsFname
-    # iterate sets of libraries
-    for i, (libnames, libFs, libRs, orientations, libIS, libISStDev) in enumerate(libraries, 1):
-        reads = ""
-        # iterate libraries withing the set
-        for j, (fq1, fq2, orient, iSize, iFrac) in enumerate(zip(libFs, libRs, orientations, \
-                                                                 libIS, libISStDev), 1):
-            # consider skipping mate-pairs is libIS>1kb
-            # skip orientations other than FR RF
-            if orient == "FR":
-                reverse_seq = 0
-            elif orient == "RF":
-                reverse_seq = 1
-            else:
-                if verbose:
-                    info = "  Skipping unsupported library %s in: %s - %s!\n"
-                    sys.stderr.write(info%(orient, fq1, fq2))
-                continue
-            # add reads
-            reads += "%s,%s" % (fq1.name, fq2.name)
-        # skip if not reads
-        if not reads:
-            continue
-        out = os.path.join(outdir, "_gap2seq.%s.%s.fa"%(i,j))
-        # run Gap2Seq
-        cmd = ["Gap2Seq", "-nb-cores %s"%threads, "-scaffolds", pout, \
-               "-filled", out, "-reads", reads]
+        out = os.path.join(outdir, "%s.%s.fa"%(basename, i))
+        # run GapCloser
+        cmd = ["GapCloser", "-t %s"%threads, "-p %s"%overlap, "-l %s"%maxReadLen, \
+               "-a", pout, "-b", configFn, "-o", out]
         if verbose:
             sys.stderr.write(" iteration %s ...\n"%i)
             #sys.stderr.write( "  %s\n" % " ".join(cmd) )
         with open(out+".log", "w") as log:
-            Gap2Seq = subprocess.Popen(cmd, bufsize=-1, stdout=log, stderr=log)
-            Gap2Seq.wait()
+            GapCloser = subprocess.Popen(cmd, bufsize=-1, stdout=log, stderr=log)
+            GapCloser.wait()
         # store out info
         pout = out
-
     # create symlink to final scaffolds or pout
     symlink(pout, nogapsFname)
-    
+        
 def redundants(fastq, fasta, outdir, mapq, threads, identity, overlap, minLength, \
                joins, readLimit, iters, sspacebin, reduction=1, scaffolding=1, gapclosing=1, \
                verbose=1, log=sys.stderr):
@@ -321,7 +294,7 @@ def redundants(fastq, fasta, outdir, mapq, threads, identity, overlap, minLength
         sys.stderr.write(fasta_stats(open(fn)))
 
     # Clean-up
-    # rm fq.is.txt
+    # rm fq.is.txt outdir/_reads,
     # rm *.h5 CHANGEME.mphf
 
 def _check_executable(cmd):
